Add copyDirectory function
